Reduce 429 errors from multiple in-flight searches (#55)
Often when I'm opening a few tabs from clicking on search results, some
will fail with a 429 error. According to the docs, this is expected more
often than we'd realised after upping `concurrent_jobs` to scale
querying in a previous PR.

From the docs:
When increasing concurrent_jobs, also increase the queue size per
tenant, or search requests will be cause 429 errors. This is the total
number of jobs per tenant allowed in the queue.